+++ src/humbug/mindspace/preview/mindspace_preview_view.py
@@ -63,7 +63,6 @@ class MindspacePreviewView(QWidget):
         # Create the three coordinated widgets and wrap them in the container.
         self._breadcrumb_bar = MindspaceBreadcrumbBar()
         self._breadcrumb_bar.set_drop_handler(self._on_file_dropped)
-        self._breadcrumb_bar.set_scroll_handler(self.reveal_and_select_file)
 
         self._tree_view = MindspacePreviewTreeView()
         self._tree_view.customContextMenuRequested.connect(self._show_context_menu)
@@ -75,6 +74,7 @@ class MindspacePreviewView(QWidget):
 
         self._bc_container = MindspaceBreadcrumbContainer(self._breadcrumb_bar, self._tree_view, self)
         layout.addWidget(self._bc_container, 1)
+        self._bc_container.set_scroll_handler(self.reveal_and_select_file)
 
         # Create file system model
         self._icon_provider = MindspaceTreeIconProvider()